Work to remove dependency on Godot Voxel Tools

+++ vis/Blocks.cs
@@ -1,6 +1,6 @@
 namespace Ritgard;
 
-public enum Blocks : ulong
+public enum Blocks : byte
 {
     Air = 0,
     Stone = 1,
